Migrate the data model around the concept of "Job"
There is no more workflow and run, a Job defines a workflow and what is
necessary to execute it. The rationale comes from our workflow, which
did not show the need to separate those two concepts, as we essentially
iterate over workflows.

- A JSON-based DSL is integrated to describe jobs, more to come.
- The infamous RunManager/SchedulerService are no more, they are
  replaced by the `accio.state` package and hopefully more flexible
  and testable classes (fixes #21).
- The storage has been improved, there is no more `write` method, and
  all its methods are now asynchronous (fixes #7).

+++ accio/java/fr/cnrs/liris/accio/sdk/Op.java
@@ -64,7 +64,7 @@ public @interface Op {
     /**
      * Number of CPU cores required to execute this operator. It will be allocated at least this number of cores.
      */
-    float cpus() default 1;
+    int cpus() default 1;
 
     /**
      * Amount of RAM required to execute this operator. It will be allocated at least this amount of RAM.
